requirements.txt
asgiref==3.6.0
cloudinary==1.33.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.18
django-allauth==0.54.0
django-bootstrap==0.2.4
django-bootstrap4==23.1
django-cloudinary-storage==0.3.0
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
